Use `file/write` to stream output
This commit uses `ev/to-file` to make the socket connection work with
`file/write`. This allows print messages to be sent to the external
client during evaluation. Asynchronous writing (such as with `ev/write`)
could not be used for this because Janet's print functions cannot call
asynchronous functions.

This fixes #1 on GitHub Issues.